Auto initialize nested editors
